feat: prisma seed 파일 위치 옮기고 수정
