+++ package.json
@@ -2,13 +2,14 @@
   "name": "book-review",
   "version": "0.1.0",
   "private": true,
-  "type": "module",
+  "prisma": {
+    "seed": "ts-node --compiler-options {\"module\":\"CommonJS\"} prisma/seed.ts"
+  },
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
-    "seed": "npx tsx scripts/seed.ts"
+    "lint": "next lint"
   },
   "dependencies": {
     "@prisma/client": "^5.17.0",
