+++ prisma/schema.prisma
@@ -144,14 +144,13 @@ model BalletPost {
 
 model BalletRecord {
   id String @id @default(dbgenerated("gen_random_uuid()")) @db.Uuid
-  month Int
-  day Int
+  date DateTime
   balletDone Boolean
   balletPosts BalletPostRecord[]
-  author User @relation(fields: [authorId], references: [id])
-  authorId String @db.Uuid
+  user User @relation(fields: [userId], references: [id])
+  userId String @db.Uuid
 
-  @@unique([month, day, authorId])
+  @@unique([date, userId])
 }
 
 model BalletPostRecord {
